Implemented read message length (need to add tests)

+++ src/Lightning/Network/Protocol/Transport/IHandshakeProtocol.cs
@@ -16,7 +16,7 @@ namespace Network.Protocol.Transport
 
       public void ReadMessage(ReadOnlySpan<byte> message, IBufferWriter<byte> output);
       
-      public long ReadMessageLength(ReadOnlySequence<byte> encryptedHeader);
+      public ushort ReadMessageLength(ReadOnlySequence<byte> encryptedHeader);
 
       public void Handshake(ReadOnlySpan<byte> message, IBufferWriter<byte> output);
    }
